+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.3.7 (to be released on 2022.07.12)
+
+* added `collection.second` as alias to `collection.elementAt(1)`
+* added `collection.by(condition)` as alias to `collection.filteredBy(condition)`
+  * filteredBy will be deprecated in future releases
+
 ## 1.3.6 (to be released on 2022.07.12)
 
 ### New Features